requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
